sampleproject/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
